pyproject.toml
[tool.poetry]
name = "gamecenter"
version = "0.1.0"
description = "Backend part of GameCenter system"
authors = ["RTU MIREA Laboratory"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
django-dotenv = "^1.4"
django-extensions = "^2.1"
django-cors-headers = "^2.4"
djangorestframework = "^3.9"
djangorestframework_simplejwt = "^3.3"
six = "^1.12"
djangorestframework-camel-case = "^1.0"
psycopg2-binary = "^2.7"
Pillow = "^5.4"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0",allows-prereleases = true}
isort = "^4.3"
pytest = "^4.3"
pytest-django = "^3.4"
pytest-cov = "^2.6"
django-extensions = "^2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"